+++ NEWS.md
@@ -9,7 +9,8 @@ Version 28.0-dev (main)
 Version 27.1-dev (v27-fixes)
 ----------------------------
 
-*
+* Workaround for a Qt bug on Android that sometimes resulted in empty
+  window if app becomes active again
 
 Version 27.0 (10 Sep 2025)
 --------------------------